drop server_sockaddr.h in favour of util.h, move the struct def there
