+++ tests/server/rtspd.c
@@ -51,7 +51,6 @@
 #include <curlx.h> /* from the private lib dir */
 #include "getpart.h"
 #include "util.h"
-#include "server_sockaddr.h"
 
 /* include memdebug.h last */
 #include <memdebug.h>
