+++ tests/server/sockfilt.c
@@ -104,7 +104,6 @@
 #include <curlx.h> /* from the private lib dir */
 #include "inet_pton.h"
 #include "util.h"
-#include "server_sockaddr.h"
 #include "timediff.h"
 
 #include "tool_binmode.h"
