+++ tests/server/tftpd.c
@@ -86,7 +86,6 @@
 #include <curlx.h> /* from the private lib dir */
 #include "getpart.h"
 #include "util.h"
-#include "server_sockaddr.h"
 
 /*****************************************************************************
 *  This is a rewrite/clone of the arpa/tftp.h file for systems without it.   *
